moved things arround, added v22 compiler for reference
Moved the 68XX compiler to it's own directory. Added the v22 (DOS)
compiler in it's own directoy. Found that the 68xx compiler compiles
to pseudo code and the I needed to do a bulk replace of int to long.
There are plenty gcc complaints of casting mismatches but long and
char * share the same 8 byte lenght (Linux = 64bit addressing) so the
cast mostly works. I'll clean up as I go.