Fixed a compiler bu would stop after reading the last byte
Found a bug where the compiler would stop after the last byte was
read. This would cause the compiler to stop it's output. I've also
changed a few function to use long instead of int. This is related to
the problem with Small C 6800 pointers (16 bit) and Linux pointers (64
bit).